+++ container-engine-lib/docs/changelog.md
@@ -1,5 +1,6 @@
 # TBD
 ### Features
+* Added `KubernetesKurtosisBackend.GetModuleLogs`
 * Added the functionality to wait until the GRPC port is available before returning when creating `Engines`, `API containers` and `Modules`  
 * Added `KubernetesKurtosisBackend.CreateModule`, `KubernetesKurtosisBackend.GetModules`, `KubernetesKurtosisBackend.StopModules` and `KubernetesKurtosisBackend.DestroyModules`
 * Added `ForModulePod` and `ForModuleService` to `KubernetesEnclaveObjectAttributesProvider`